[ edit ] auth controller interface
