--- document/directory.md
@@ -1,2 +1,0 @@
-## 디렉터리 구조 설명
-
