+++ src/domain/user/infrastructure/entities/enums/user.enum.role.ts
@@ -1,0 +1,7 @@
+export enum UserRole {
+  UNKNOWN   = 'UNKNOWN',
+  ADMIN     = 'ADMIN',
+  CLIENT    = 'CLIENT',
+  SELLER    = 'SELLER',
+  PUBLISHER = 'PUBLISHER',
+}
