+++ src/lib/authentication/presentation/interface/auth.controller.interface.ts
@@ -1,9 +1,14 @@
+import type { Request, Response }   from "express";
 import { RegisterUserDto } from "../dto/auth.register.user.dto";
+import { LoginDto }        from "../dto/login.firebase.user.dto";
+import { IToken }          from "./token.interface";
 
 
 
-export interface IAuthController<T> {
+export interface IAuthController {
   registerUserThroughAuthenticationServer( registerUserDto: RegisterUserDto ): Promise<any>;
 
-  // registerUserThroughCognito( cognitoUserDto ): Promise<any>
+  login( loginDto: LoginDto, res: Response ): Promise<IToken>;
+
+  refresh( req: Request ): Promise<Pick<IToken, "access_token">>;
 }
